feat: add small evolutionary design project structure
- created 1-init/project-structure/ folder in src
- added base folders: app, features, shared
- app: App.tsx, main.tsx, index.css (global styles)
- features:
  - auth: AuthPage, RegistrationPage
  - board: BoardPage
  - board-list: BoardListPage
  - header: AppHeaderPage
