+++ src/1-init/project-structure/app/App.tsx
@@ -1,0 +1,9 @@
+function App() {
+  return (
+    <>
+      <div>Hello world</div>
+    </>
+  );
+}
+
+export default App;
