+++ src/1-init/project-structure/app/index.css
@@ -1,0 +1,3 @@
+* {
+  color: red;
+}
