+++ src/1-init/project-structure/feature/auth/index.tsx
@@ -1,0 +1,17 @@
+export const LoginPage = () => {
+  return (
+    <div>
+      <h1>Login Page</h1>
+      <p>Please enter your credentials to log in.</p>
+    </div>
+  );
+};
+
+export const RegisterPage = () => {
+  return (
+    <div>
+      <h1>Register Page</h1>
+      <p>Please fill in the form to create an account.</p>
+    </div>
+  );
+};
