+++ src/1-init/project-structure/feature/board-list/index.tsx
@@ -1,0 +1,7 @@
+export const BoardsListPage = () => {
+  return (
+    <>
+      <div>Boards List Page</div>;
+    </>
+  );
+};
