+++ src/1-init/project-structure/feature/board/index.tsx
@@ -1,0 +1,7 @@
+export const BoardPage = () => {
+  return (
+    <>
+      <div>Board Page</div>
+    </>
+  );
+};
