+++ src/1-init/project-structure/feature/header/index.tsx
@@ -1,0 +1,7 @@
+export const AppHeader = () => {
+  return (
+    <header>
+      <h1>My App Header</h1>
+    </header>
+  );
+};
