+++ src/1-init/project-structure/shared/vite-env.d.ts
