--- src/App.tsx
@@ -1,5 +1,0 @@
-function App() {
-  return <></>;
-}
-
-export default App;
